Separated out EventDrivenTrajectoryService
